+++ extensions/qqbot/src/config.ts
@@ -1,98 +1,30 @@
 import fs from "node:fs";
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-runtime";
 import { getPlatformAdapter } from "./engine/adapter/index.js";
-import { normalizeOptionalLowercaseString } from "./engine/utils/string-normalize.js";
+import {
+  DEFAULT_ACCOUNT_ID as ENGINE_DEFAULT_ACCOUNT_ID,
+  listAccountIds,
+  resolveAccountBase,
+  resolveDefaultAccountId,
+} from "./engine/config/resolve.js";
 import type { ResolvedQQBotAccount, QQBotAccountConfig } from "./types.js";
 
-export const DEFAULT_ACCOUNT_ID = "default";
+export const DEFAULT_ACCOUNT_ID = ENGINE_DEFAULT_ACCOUNT_ID;
 
 interface QQBotChannelConfig extends QQBotAccountConfig {
   accounts?: Record<string, QQBotAccountConfig>;
   defaultAccount?: string;
 }
 
-function normalizeConfiguredDefaultAccountId(raw: unknown): string | null {
-  return normalizeOptionalLowercaseString(raw) ?? null;
-}
-
-function normalizeQQBotAccountConfig(account: QQBotAccountConfig | undefined): QQBotAccountConfig {
-  if (!account) {
-    return {};
-  }
-  return {
-    ...account,
-    ...(account.audioFormatPolicy ? { audioFormatPolicy: { ...account.audioFormatPolicy } } : {}),
-  };
-}
-
-function normalizeAppId(raw: unknown): string {
-  if (typeof raw === "string") {
-    return raw.trim();
-  }
-  if (typeof raw === "number") {
-    return String(raw);
-  }
-  return "";
-}
-
-function buildQQBotAccountConfigPatch(input: {
-  appId?: string;
-  clientSecret?: string;
-  clientSecretFile?: string;
-  name?: string;
-}): Partial<QQBotAccountConfig> {
-  return {
-    ...(input.appId ? { appId: input.appId } : {}),
-    ...(input.clientSecret
-      ? { clientSecret: input.clientSecret, clientSecretFile: undefined }
-      : input.clientSecretFile
-        ? { clientSecretFile: input.clientSecretFile, clientSecret: undefined }
-        : {}),
-    ...(input.name ? { name: input.name } : {}),
-  };
-}
 
 /** List all configured QQBot account IDs. */
 export function listQQBotAccountIds(cfg: OpenClawConfig): string[] {
-  const ids = new Set<string>();
-  const qqbot = cfg.channels?.qqbot as QQBotChannelConfig | undefined;
-
-  if (qqbot?.appId || process.env.QQBOT_APP_ID) {
-    ids.add(DEFAULT_ACCOUNT_ID);
-  }
-
-  if (qqbot?.accounts) {
-    for (const accountId of Object.keys(qqbot.accounts)) {
-      if (qqbot.accounts[accountId]?.appId) {
-        ids.add(accountId);
-      }
-    }
-  }
-
-  return Array.from(ids);
+  return listAccountIds(cfg as unknown as Record<string, unknown>);
 }
 
 /** Resolve the default QQBot account ID. */
 export function resolveDefaultQQBotAccountId(cfg: OpenClawConfig): string {
-  const qqbot = cfg.channels?.qqbot as QQBotChannelConfig | undefined;
-  const configuredDefaultAccountId = normalizeConfiguredDefaultAccountId(qqbot?.defaultAccount);
-  if (
-    configuredDefaultAccountId &&
-    (configuredDefaultAccountId === DEFAULT_ACCOUNT_ID ||
-      Boolean(qqbot?.accounts?.[configuredDefaultAccountId]?.appId))
-  ) {
-    return configuredDefaultAccountId;
-  }
-  if (qqbot?.appId || process.env.QQBOT_APP_ID) {
-    return DEFAULT_ACCOUNT_ID;
-  }
-  if (qqbot?.accounts) {
-    const ids = Object.keys(qqbot.accounts);
-    if (ids.length > 0) {
-      return ids[0];
-    }
-  }
-  return DEFAULT_ACCOUNT_ID;
+  return resolveDefaultAccountId(cfg as unknown as Record<string, unknown>);
 }
 
 /** Resolve QQBot account config for runtime or setup flows. */
@@ -101,31 +33,23 @@ export function resolveQQBotAccount(
   accountId?: string | null,
   opts?: { allowUnresolvedSecretRef?: boolean },
 ): ResolvedQQBotAccount {
-  const resolvedAccountId = accountId ?? resolveDefaultQQBotAccountId(cfg);
-  const qqbot = cfg.channels?.qqbot as QQBotChannelConfig | undefined;
+  const raw = cfg as unknown as Record<string, unknown>;
+  const base = resolveAccountBase(raw, accountId);
 
-  let accountConfig: QQBotAccountConfig = {};
-  let appId = "";
+  const qqbot = cfg.channels?.qqbot as QQBotChannelConfig | undefined;
+  const accountConfig: QQBotAccountConfig =
+    base.accountId === DEFAULT_ACCOUNT_ID
+      ? (qqbot ?? {})
+      : (qqbot?.accounts?.[base.accountId] ?? {});
+
   let clientSecret = "";
   let secretSource: "config" | "file" | "env" | "none" = "none";
 
-  if (resolvedAccountId === DEFAULT_ACCOUNT_ID) {
-    // Default account reads from top-level config and keeps the full field surface.
-    accountConfig = normalizeQQBotAccountConfig(qqbot);
-    appId = normalizeAppId(qqbot?.appId);
-  } else {
-    // Named accounts read from channels.qqbot.accounts.
-    const account = qqbot?.accounts?.[resolvedAccountId];
-    accountConfig = normalizeQQBotAccountConfig(account);
-    appId = normalizeAppId(account?.appId);
-  }
-
   const clientSecretPath =
-    resolvedAccountId === DEFAULT_ACCOUNT_ID
+    base.accountId === DEFAULT_ACCOUNT_ID
       ? "channels.qqbot.clientSecret"
-      : `channels.qqbot.accounts.${resolvedAccountId}.clientSecret`;
+      : `channels.qqbot.accounts.${base.accountId}.clientSecret`;
 
-  // Resolve clientSecret from config, file, or environment.
   const adapter = getPlatformAdapter();
   if (adapter.hasConfiguredSecret(accountConfig.clientSecret)) {
     clientSecret = opts?.allowUnresolvedSecretRef
@@ -142,25 +66,20 @@ export function resolveQQBotAccount(
     } catch {
       secretSource = "none";
     }
-  } else if (process.env.QQBOT_CLIENT_SECRET && resolvedAccountId === DEFAULT_ACCOUNT_ID) {
+  } else if (process.env.QQBOT_CLIENT_SECRET && base.accountId === DEFAULT_ACCOUNT_ID) {
     clientSecret = process.env.QQBOT_CLIENT_SECRET;
     secretSource = "env";
   }
 
-  // AppId can also fall back to an environment variable.
-  if (!appId && process.env.QQBOT_APP_ID && resolvedAccountId === DEFAULT_ACCOUNT_ID) {
-    appId = normalizeAppId(process.env.QQBOT_APP_ID);
-  }
-
   return {
-    accountId: resolvedAccountId,
+    accountId: base.accountId,
     name: accountConfig.name,
-    enabled: accountConfig.enabled !== false,
-    appId,
+    enabled: base.enabled,
+    appId: base.appId,
     clientSecret,
     secretSource,
-    systemPrompt: accountConfig.systemPrompt,
-    markdownSupport: accountConfig.markdownSupport !== false,
+    systemPrompt: base.systemPrompt,
+    markdownSupport: base.markdownSupport,
     config: accountConfig,
   };
 }
@@ -177,10 +96,8 @@ export function applyQQBotAccountConfig(
   },
 ): OpenClawConfig {
   const next = { ...cfg };
-  const accountConfigPatch = buildQQBotAccountConfigPatch(input);
 
   if (accountId === DEFAULT_ACCOUNT_ID) {
-    // Default allowFrom to ["*"] when not yet configured.
     const existingConfig = (next.channels?.qqbot as QQBotChannelConfig) || {};
     const allowFrom = existingConfig.allowFrom ?? ["*"];
 
@@ -190,11 +107,16 @@ export function applyQQBotAccountConfig(
         ...(next.channels?.qqbot as Record<string, unknown> | undefined),
         enabled: true,
         allowFrom,
-        ...accountConfigPatch,
+        ...(input.appId ? { appId: input.appId } : {}),
+        ...(input.clientSecret
+          ? { clientSecret: input.clientSecret, clientSecretFile: undefined }
+          : input.clientSecretFile
+            ? { clientSecretFile: input.clientSecretFile, clientSecret: undefined }
+            : {}),
+        ...(input.name ? { name: input.name } : {}),
       },
     };
   } else {
-    // Default allowFrom to ["*"] when not yet configured.
     const existingAccountConfig =
       (next.channels?.qqbot as QQBotChannelConfig)?.accounts?.[accountId] || {};
     const allowFrom = existingAccountConfig.allowFrom ?? ["*"];
@@ -210,7 +132,13 @@ export function applyQQBotAccountConfig(
             ...(next.channels?.qqbot as QQBotChannelConfig)?.accounts?.[accountId],
             enabled: true,
             allowFrom,
-            ...accountConfigPatch,
+            ...(input.appId ? { appId: input.appId } : {}),
+            ...(input.clientSecret
+              ? { clientSecret: input.clientSecret, clientSecretFile: undefined }
+              : input.clientSecretFile
+                ? { clientSecretFile: input.clientSecretFile, clientSecret: undefined }
+                : {}),
+            ...(input.name ? { name: input.name } : {}),
           },
         },
       },
