+++ extensions/qqbot/src/engine/ref/store.ts
@@ -2,12 +2,12 @@
  * Ref-index store — JSONL file-based store for message reference index.
  *
  * Migrated from src/ref-index-store.ts. Dependencies are only Node.js
- * built-ins + debug-log + platform (both zero plugin-sdk).
+ * built-ins + log + platform (both zero plugin-sdk).
  */
 
 import fs from "node:fs";
 import path from "node:path";
-import { debugLog, debugError } from "../utils/debug-log.js";
+import { debugLog, debugError } from "../utils/log.js";
 import { getQQBotDataDir } from "../utils/platform.js";
 import type { RefIndexEntry } from "./types.js";
